+++ src/components/UI/ButtonSelect/styles.ts
@@ -1,13 +1,13 @@
 import styled from "styled-components/native";
 import { theme } from "../../../styles/theme";
-import { TouchableOpacity } from "react-native";
+import { TouchableOpacity } from "react-native-gesture-handler";
 
-
-export const Wrapper = styled(TouchableOpacity)<{color: string}>`
+export const WrapperSelect = styled(TouchableOpacity)<{color: string}>`
     width: 100%;
     height: 100%;
     align-items: center;
-    border-radius: 6px;
     justify-content: center;
+    border-radius: 160px;
+    border:${theme.border.line}px ${theme.border.color} ;
     background-color: ${(props) => theme.colors[props.color]};
 `