+++ src/components/UI/Input/styles.ts
@@ -5,9 +5,9 @@ import { theme } from "../../../styles/theme";
 export const InputWrapper = styled.TextInput`
     width: 100%;
     height: 45px;
-    border: ${theme.border.line}px solid ${theme.border.color};
+    border: 1px solid ${theme.border.color};
     border-radius: 6px;
     padding: 8px;
     font-size: 14px;
-    margin-bottom: 4%;
+    background-color: ${theme.colors.backgroundWhite};
 `