+++ src/routes/AppRoutes.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
 import { createStackNavigator } from "@react-navigation/stack";
 import { ThemeProvider } from '@react-navigation/native';
-import Home from '../screens/Home';
+import Welcome from '../screens/Welcome';
 import SignIn from '../screens/SignIn';
 import { theme } from '../styles/theme';
 import SignUp from '../screens/SignUp';
-import HomeScreen from '../screens/HomeScreen';
+import Home from '../screens/Home';
 
 export type StackPramsList = {
-  Home: undefined;
+  Welcome: undefined;
   SignIn: undefined;
   SignUp: undefined;
-  HomeScreen: {
-    email: string;
+  Home: {
+    userEmail: string;
   };
 }
 const Stack = createStackNavigator<StackPramsList>();
@@ -25,10 +25,10 @@ export default function Routes() {
   return (
     <ThemeProvider value={theme}>
       <Stack.Navigator>
-        <Stack.Screen name="Home" component={Home} options={routeOptions} />
+        <Stack.Screen name="Welcome" component={Welcome} options={routeOptions} />
         <Stack.Screen name="SignIn" component={SignIn} options={routeOptions} />
         <Stack.Screen name='SignUp' component={SignUp} options={routeOptions} />
-        <Stack.Screen name='HomeScreen' component={HomeScreen } options={routeOptions} />
+        <Stack.Screen name='Home' component={Home} options={routeOptions} />
       </Stack.Navigator>
     </ThemeProvider>
   );